Fix signedness warnings in cpp tests.
The compiler infers the type of unsuffixed literals as signed. That causes GCC
warnings like this with gtest's ASSERT_EQ if one side is a literal and other is
an unsigned type:

In file included from src/test/cpp/option_processor_test.cc:23:0:
external/com_google_googletest/googletest/include/gtest/gtest.h: In instantiation of 'testing::AssertionResult testing::internal::CmpHelperEQ(const char*, const char*, const T1&, const T2&) [with T1 = int; T2 = long unsigned int]':
external/com_google_googletest/googletest/include/gtest/gtest.h:1449:23:   required from 'static testing::AssertionResult testing::internal::EqHelper<lhs_is_null_literal>::Compare(const char*, const char*, const T1&, const T2&) [with T1 = int; T2 = long unsigned int; bool lhs_is_null_literal = false]'
src/test/cpp/option_processor_test.cc:98:3:   required from here
external/com_google_googletest/googletest/include/gtest/gtest.h:1421:11: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
   if (lhs == rhs) {
       ~~~~^~~~~~

+++ src/test/cpp/option_processor_test.cc
@@ -95,7 +95,7 @@ TEST_F(OptionProcessorTest, CanParseOptions) {
                 << error;
 
   ASSERT_EQ("", error);
-  ASSERT_EQ(1,
+  ASSERT_EQ(size_t(1),
             option_processor_->GetParsedStartupOptions()->host_jvm_args.size());
   EXPECT_EQ("MyParam",
             option_processor_->GetParsedStartupOptions()->host_jvm_args[0]);
@@ -119,7 +119,7 @@ TEST_F(OptionProcessorTest, CanParseHelpArgs) {
                 << error;
 
   ASSERT_EQ("", error);
-  ASSERT_EQ(1,
+  ASSERT_EQ(size_t(1),
             option_processor_->GetParsedStartupOptions()->host_jvm_args.size());
   EXPECT_EQ("MyParam",
             option_processor_->GetParsedStartupOptions()->host_jvm_args[0]);
@@ -155,7 +155,7 @@ TEST_F(OptionProcessorTest, CanParseDifferentStartupArgs) {
                 << error;
   ASSERT_EQ("", error);
 
-  ASSERT_EQ(2,
+  ASSERT_EQ(size_t(2),
             option_processor_->GetParsedStartupOptions()->host_jvm_args.size());
   EXPECT_EQ("MyParam",
             option_processor_->GetParsedStartupOptions()->host_jvm_args[0]);
